requirements.txt
pandas
dune-client
scikit-learn
python-dotenv
matplotlib
seaborn
requests